requirements.txt
openai
python-dotenv
colorama
tiktoken